requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.5
django-braces==1.15.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
wrapt==1.15.0
